sportsanalytics-dashboard-main/requirements.txt
dash==2.9.3
pandas==1.4.4
plotly~=5.14.1
dash-bootstrap-components==1.4.1
python-dotenv==0.19.2
sweat==0.25.0
urllib3~=2.0.3
dash-leaflet~=0.1.23
fitparse~=1.2.0
numpy~=1.24.3
dash_svg==0.0.12